requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0

# ── Database ──────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
psycopg2-binary>=2.9.0

# ── Auth & Security ──────────────────────────────────────────────────────
bcrypt>=4.2.0
PyJWT>=2.9.0
itsdangerous>=2.2.0

# ── Redis ─────────────────────────────────────────────────────────────────
redis>=5.2.0

# ── RabbitMQ ──────────────────────────────────────────────────────────────
aio-pika>=9.5.0

# ── Scheduler ─────────────────────────────────────────────────────────────
APScheduler>=3.10.0

# ── Admin ─────────────────────────────────────────────────────────────────
sqladmin>=0.20.0

# ── Observability ─────────────────────────────────────────────────────────
sentry-sdk[fastapi]>=2.19.0

# ── HTTP client ───────────────────────────────────────────────────────────
httpx>=0.28.0

# ── CLI ───────────────────────────────────────────────────────────────────
typer>=0.15.0
rich>=13.9.0
